+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added CMHeartbeatDiscovery module
 - Added ConvertTo-ScheduleInterval to ResourceHelper
 - Added CMServiceConnectionPoint Resource
+- Added CMNetworkDiscovery Resource
 
 ### Changed